requirements.txt
databricks-sql-connector[pyarrow]
openai
python-dotenv
fastapi
uvicorn
pyyaml